+++ packages/database/src/outbox/models/outbox.model.ts
@@ -1,0 +1,15 @@
+import type { OutboxModelStatus } from '../types/outbox-model.status.js';
+
+export abstract class OutboxModel {
+  id!: string;
+
+  // Tracking and attempts
+  status!: OutboxModelStatus;
+  attemps!: number;
+  lastError?: string;
+
+  type!: string;
+  emitter!: string;
+
+  createdAt!: Date;
+}