+++ packages/domain-event/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.2.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/database@1.4.0
+
 ## 2.2.9
 
 ### Patch Changes